app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pulsar-client==3.5.0
